Revert "feat: read qrcode from image (#774)"
This reverts commit f07be0f592af61412b2444e32bd8a5d4b9f1866b.
