+++ apps/mobile/android/build.gradle
@@ -13,7 +13,6 @@ buildscript {
         supportLibVersion = "34.0.0"
         googlePlayServicesAuthVersion = "20.7.0"
         kotlin_version = '1.9.24'
-        kotlinVersion = kotlin_version
     }
     repositories {
         google()
