+++ apps/mobile/ios/PrivacyInfo.xcprivacy
@@ -20,7 +20,6 @@
 			<key>NSPrivacyAccessedAPITypeReasons</key>
 			<array>
 				<string>C617.1</string>
-				<string>3B52.1</string>
 			</array>
 		</dict>
 		<dict>
