+++ apps/mobile/package.json
@@ -137,7 +137,6 @@
     "react-native": "0.73.11",
     "react-native-aes-crypto": "^3.0.1",
     "react-native-animateable-text": "0.11.1",
-    "react-native-barcodes-detector": "^0.1.1",
     "react-native-ble-manager": "8",
     "react-native-ble-plx": "3.1.2",
     "react-native-bundle-splitter": "^3.0.1",
@@ -153,7 +152,6 @@
     "react-native-get-random-values": "^1.10.0",
     "react-native-haptic-feedback": "2.3.3",
     "react-native-iap": "12.16.2",
-    "react-native-image-picker": "^8.2.0",
     "react-native-inappbrowser-reborn": "^3.7.0",
     "react-native-indicators": "^0.17.0",
     "react-native-ios-context-menu": "1.15.3",
